[APP-799] package update grantee & granter
