+++ app/src/main/java/wannabit/io/cosmostaion/activities/txs/authz/AuthzClaimCommissionActivity.java
@@ -31,8 +31,8 @@ import wannabit.io.cosmostaion.base.BaseFragment;
 import wannabit.io.cosmostaion.base.chains.ChainFactory;
 import wannabit.io.cosmostaion.fragment.StepFeeSetFragment;
 import wannabit.io.cosmostaion.fragment.StepMemoFragment;
-import wannabit.io.cosmostaion.fragment.txs.authz.grantee.AuthzClaimCommissionStep0Fragment;
-import wannabit.io.cosmostaion.fragment.txs.authz.grantee.AuthzClaimCommissionStep3Fragment;
+import wannabit.io.cosmostaion.fragment.txs.authz.granter.AuthzClaimCommissionStep0Fragment;
+import wannabit.io.cosmostaion.fragment.txs.authz.granter.AuthzClaimCommissionStep3Fragment;
 import wannabit.io.cosmostaion.model.type.Coin;
 import wannabit.io.cosmostaion.task.gRpcTask.WithdrawAddressGrpcTask;
 import wannabit.io.cosmostaion.task.gRpcTask.broadcast.AuthzClaimCommissionGrpcTask;
