+++ app/src/main/java/wannabit/io/cosmostaion/activities/txs/authz/AuthzClaimRewardActivity.java
@@ -32,8 +32,8 @@ import wannabit.io.cosmostaion.base.BaseFragment;
 import wannabit.io.cosmostaion.base.chains.ChainFactory;
 import wannabit.io.cosmostaion.fragment.StepFeeSetFragment;
 import wannabit.io.cosmostaion.fragment.StepMemoFragment;
-import wannabit.io.cosmostaion.fragment.txs.authz.grantee.AuthzClaimRewardStep0Fragment;
-import wannabit.io.cosmostaion.fragment.txs.authz.grantee.AuthzClaimRewardStep3Fragment;
+import wannabit.io.cosmostaion.fragment.txs.authz.granter.AuthzClaimRewardStep0Fragment;
+import wannabit.io.cosmostaion.fragment.txs.authz.granter.AuthzClaimRewardStep3Fragment;
 import wannabit.io.cosmostaion.model.type.Coin;
 import wannabit.io.cosmostaion.task.gRpcTask.WithdrawAddressGrpcTask;
 import wannabit.io.cosmostaion.task.gRpcTask.broadcast.AuthzClaimRewardGrpcTask;
