+++ app/src/main/java/wannabit/io/cosmostaion/activities/txs/authz/AuthzSendActivity.java
@@ -33,9 +33,9 @@ import wannabit.io.cosmostaion.base.BaseFragment;
 import wannabit.io.cosmostaion.base.chains.ChainFactory;
 import wannabit.io.cosmostaion.fragment.StepFeeSetFragment;
 import wannabit.io.cosmostaion.fragment.StepMemoFragment;
-import wannabit.io.cosmostaion.fragment.txs.authz.grantee.AuthzSendStep0Fragment;
-import wannabit.io.cosmostaion.fragment.txs.authz.grantee.AuthzSendStep1Fragment;
-import wannabit.io.cosmostaion.fragment.txs.authz.grantee.AuthzSendStep4Fragment;
+import wannabit.io.cosmostaion.fragment.txs.authz.granter.AuthzSendStep0Fragment;
+import wannabit.io.cosmostaion.fragment.txs.authz.granter.AuthzSendStep1Fragment;
+import wannabit.io.cosmostaion.fragment.txs.authz.granter.AuthzSendStep4Fragment;
 import wannabit.io.cosmostaion.model.type.Coin;
 import wannabit.io.cosmostaion.task.gRpcTask.broadcast.AuthzSendGrpcTask;
