+++ app/src/main/java/wannabit/io/cosmostaion/activities/txs/authz/AuthzVoteActivity.java
@@ -34,9 +34,9 @@ import wannabit.io.cosmostaion.base.BaseFragment;
 import wannabit.io.cosmostaion.base.chains.ChainFactory;
 import wannabit.io.cosmostaion.fragment.StepFeeSetFragment;
 import wannabit.io.cosmostaion.fragment.StepMemoFragment;
-import wannabit.io.cosmostaion.fragment.txs.authz.grantee.AuthzVoteStep0Fragment;
-import wannabit.io.cosmostaion.fragment.txs.authz.grantee.AuthzVoteStep1Fragment;
-import wannabit.io.cosmostaion.fragment.txs.authz.grantee.AuthzVoteStep4Fragment;
+import wannabit.io.cosmostaion.fragment.txs.authz.granter.AuthzVoteStep0Fragment;
+import wannabit.io.cosmostaion.fragment.txs.authz.granter.AuthzVoteStep1Fragment;
+import wannabit.io.cosmostaion.fragment.txs.authz.granter.AuthzVoteStep4Fragment;
 import wannabit.io.cosmostaion.network.res.ResV1Proposal;
 import wannabit.io.cosmostaion.task.gRpcTask.broadcast.AuthzVoteGrpcTask;
