+++ app/src/main/java/wannabit/io/cosmostaion/fragment/txs/authz/granter/AuthzClaimRewardStep0Fragment.java
@@ -1,4 +1,4 @@
-package wannabit.io.cosmostaion.fragment.txs.authz.grantee;
+package wannabit.io.cosmostaion.fragment.txs.authz.granter;
 
 import android.os.Bundle;
 import android.text.TextUtils;
@@ -7,9 +7,12 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.LinearLayout;
+import android.widget.RelativeLayout;
 import android.widget.TextView;
 
 import androidx.annotation.Nullable;
+
+import java.util.ArrayList;
 
 import cosmos.distribution.v1beta1.Distribution;
 import cosmos.staking.v1beta1.Staking;
@@ -19,18 +22,17 @@ import wannabit.io.cosmostaion.base.BaseFragment;
 import wannabit.io.cosmostaion.model.type.Coin;
 import wannabit.io.cosmostaion.utils.WDp;
 
-public class AuthzClaimRewardStep3Fragment extends BaseFragment implements View.OnClickListener {
+public class AuthzClaimRewardStep0Fragment extends BaseFragment implements View.OnClickListener {
 
-    private TextView mTvRewardAmount;
-    private TextView mFeeAmount;
+    private TextView mTvRewardAmount, mTvDenomTitle;
     private TextView mTvFromValidators;
-    private LinearLayout mTvGoalLayer;
-    private TextView mTvGoalAddress, mMemo;
-    private Button mBeforeBtn, mConfirmBtn;
-    private TextView mRewardDenom, mFeeDenom;
+    private LinearLayout mReceiveLayer;
+    private TextView mTvReceiveAddress;
+    private RelativeLayout mProgressBar;
+    private Button mCancelBtn, mNextBtn;
 
-    public static AuthzClaimRewardStep3Fragment newInstance() {
-        return new AuthzClaimRewardStep3Fragment();
+    public static AuthzClaimRewardStep0Fragment newInstance() {
+        return new AuthzClaimRewardStep0Fragment();
     }
 
     @Override
@@ -40,30 +42,28 @@ public class AuthzClaimRewardStep3Fragment extends BaseFragment implements View.
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        View rootView = inflater.inflate(R.layout.fragment_authz_reward_step3, container, false);
-        mTvRewardAmount = rootView.findViewById(R.id.reward_amount);
-        mFeeAmount = rootView.findViewById(R.id.reward_fees);
-        mTvFromValidators = rootView.findViewById(R.id.reward_moniker);
-        mTvGoalLayer = rootView.findViewById(R.id.reward_receive_layer);
-        mTvGoalAddress = rootView.findViewById(R.id.reward_receive_address);
-        mMemo = rootView.findViewById(R.id.memo);
-        mRewardDenom = rootView.findViewById(R.id.reward_denom);
-        mFeeDenom = rootView.findViewById(R.id.reward_fees_type);
-        mBeforeBtn = rootView.findViewById(R.id.btn_before);
-        mConfirmBtn = rootView.findViewById(R.id.btn_confirm);
+        View rootView = inflater.inflate(R.layout.fragment_authz_reward_step0, container, false);
+        mTvRewardAmount         = rootView.findViewById(R.id.reward_amount);
+        mTvDenomTitle           = rootView.findViewById(R.id.reward_denom);
+        mTvFromValidators       = rootView.findViewById(R.id.reward_moniker);
+        mReceiveLayer           = rootView.findViewById(R.id.reward_receive_address_layer);
+        mTvReceiveAddress       = rootView.findViewById(R.id.reward_receive_address);
+        mProgressBar            = rootView.findViewById(R.id.reward_progress);
+        mCancelBtn              = rootView.findViewById(R.id.btn_cancel);
+        mNextBtn                = rootView.findViewById(R.id.btn_next);
 
-        mBeforeBtn.setOnClickListener(this);
-        mConfirmBtn.setOnClickListener(this);
+        mCancelBtn.setOnClickListener(this);
+        mNextBtn.setOnClickListener(this);
         return rootView;
     }
 
     @Override
     public void onRefreshTab() {
         Coin mainReward = getSActivity().mGranterRewardSum;
-        WDp.setDpCoin(getSActivity(), getBaseDao(), getSActivity().mChainConfig, mainReward, mRewardDenom, mTvRewardAmount);
-        WDp.setDpCoin(getSActivity(), getBaseDao(), getSActivity().mChainConfig, getSActivity().mTxFee.amount.get(0), mFeeDenom, mFeeAmount);
+        WDp.setDpCoin(getActivity(), getBaseDao(), getSActivity().mChainConfig, mainReward, mTvDenomTitle, mTvRewardAmount);
 
         String monikers = "";
+        ArrayList<String> myValiatorList = new ArrayList<>();
         for (Staking.Validator validator: getBaseDao().mGRpcAllValidators) {
             for (Distribution.DelegationDelegatorReward myValidator : getSActivity().mGranterRewards) {
                 if (validator.getOperatorAddress().equalsIgnoreCase(myValidator.getValidatorAddress())) {
@@ -72,28 +72,30 @@ public class AuthzClaimRewardStep3Fragment extends BaseFragment implements View.
                     } else {
                         monikers = validator.getDescription().getMoniker();
                     }
+                    myValiatorList.add(myValidator.getValidatorAddress());
                 }
+                getSActivity().mValAddresses = myValiatorList;
             }
         }
         mTvFromValidators.setText(monikers);
 
-        mTvGoalAddress.setText(getSActivity().mWithdrawAddress);
+        mTvReceiveAddress.setText(getSActivity().mWithdrawAddress);
         if (getSActivity().mGranter.equalsIgnoreCase(getSActivity().mWithdrawAddress)) {
-            mTvGoalLayer.setVisibility(View.GONE);
+            mReceiveLayer.setVisibility(View.GONE);
         } else {
-            mTvGoalLayer.setVisibility(View.VISIBLE);
+            mReceiveLayer.setVisibility(View.VISIBLE);
         }
-
-        mMemo.setText(getSActivity().mTxMemo);
+        mProgressBar.setVisibility(View.GONE);
     }
+
 
     @Override
     public void onClick(View v) {
-        if (v.equals(mBeforeBtn)) {
+        if (v.equals(mCancelBtn)) {
             getSActivity().onBeforeStep();
 
-        } else if (v.equals(mConfirmBtn)) {
-            getSActivity().onAuthzClaimReward();
+        } else if (v.equals(mNextBtn)) {
+            getSActivity().onNextStep();
         }
     }
