+++ app/src/main/java/wannabit/io/cosmostaion/fragment/txs/authz/granter/AuthzDelegateStep1Fragment.java
@@ -1,4 +1,4 @@
-package wannabit.io.cosmostaion.fragment.txs.authz.grantee;
+package wannabit.io.cosmostaion.fragment.txs.authz.granter;
 
 import android.os.Bundle;
 import android.text.Editable;
@@ -21,14 +21,13 @@ import java.math.BigDecimal;
 import java.math.RoundingMode;
 
 import cosmos.staking.v1beta1.Authz;
-import cosmos.staking.v1beta1.Staking;
 import wannabit.io.cosmostaion.R;
-import wannabit.io.cosmostaion.activities.txs.authz.AuthzUndelegateActivity;
+import wannabit.io.cosmostaion.activities.txs.authz.AuthzDelegateActivity;
 import wannabit.io.cosmostaion.base.BaseFragment;
 import wannabit.io.cosmostaion.model.type.Coin;
 import wannabit.io.cosmostaion.utils.WDp;
 
-public class AuthzUndelegateStep1Fragment extends BaseFragment implements View.OnClickListener {
+public class AuthzDelegateStep1Fragment extends BaseFragment implements View.OnClickListener {
 
     private Button mCancel, mNextBtn;
     private EditText mAmountInput;
@@ -37,17 +36,19 @@ public class AuthzUndelegateStep1Fragment extends BaseFragment implements View.O
     private ImageView mClearAll;
     private Button mAdd01, mAdd1, mAdd10, mAdd100, mAddHalf, mAddMax;
 
-    private BigDecimal mGranterUndelegatable = BigDecimal.ZERO;
+    private BigDecimal mAvailable = BigDecimal.ZERO;
+    private BigDecimal mVesting = BigDecimal.ZERO;
+    private BigDecimal mGranterDelegatable = BigDecimal.ZERO;
     private int mDpDecimal = 6;
     private String mDecimalChecker, mDecimalSetter;
 
-    public static AuthzUndelegateStep1Fragment newInstance() {
-        return new AuthzUndelegateStep1Fragment();
+    public static AuthzDelegateStep1Fragment newInstance() {
+        return new AuthzDelegateStep1Fragment();
     }
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        View rootView = inflater.inflate(R.layout.fragment_authz_undelegate_step1, container, false);
+        View rootView = inflater.inflate(R.layout.fragment_authz_delegate_step1, container, false);
         mCancel = rootView.findViewById(R.id.btn_cancel);
         mNextBtn = rootView.findViewById(R.id.btn_next);
         mAmountInput = rootView.findViewById(R.id.et_amount_coin);
@@ -74,32 +75,39 @@ public class AuthzUndelegateStep1Fragment extends BaseFragment implements View.O
     }
 
     @Override
-    public void onRefreshTab() {
+    public void onResume() {
+        super.onResume();
         if (!isAdded() || getSActivity() == null || getSActivity().mAccount == null)
             getSActivity().onBackPressed();
         mDpDecimal = WDp.getDenomDecimal(getBaseDao(), getSActivity().mChainConfig, getSActivity().mChainConfig.mainDenom());
         setDpDecimals(mDpDecimal);
 
-        String selectedValAddress = getSActivity().mValAddress;
-        if (getSActivity().mGranterDelegations != null && getSActivity().mGranterDelegations.size() > 0) {
-            Staking.DelegationResponse delegated = getSActivity().mGranterDelegations.stream().filter(item -> item.getDelegation().getValidatorAddress().equalsIgnoreCase(selectedValAddress)).findFirst().get();
-            mGranterUndelegatable = new BigDecimal(delegated.getBalance().getAmount());
+        if (getSActivity().mGrantAvailable != null && getSActivity().mGrantAvailable.size() > 0) {
+            Coin availableCoin = getSActivity().mGrantAvailable.stream().filter(item -> item.denom.equalsIgnoreCase(getSActivity().mChainConfig.mainDenom())).findFirst().get();
+            mAvailable = new BigDecimal(availableCoin.amount);
         }
+        if (getSActivity().mGrantVesting != null && getSActivity().mGrantVesting.size() > 0) {
+            Coin vestingCoin = getSActivity().mGrantVesting.stream().filter(item -> item.denom.equalsIgnoreCase(getSActivity().mChainConfig.mainDenom())).findFirst().get();
+            mVesting = new BigDecimal(vestingCoin.amount);
+        }
+
+        mGranterDelegatable = mAvailable.add(mVesting);
+        mAvailableAmount.setText(WDp.getDpAmount2(getContext(), mGranterDelegatable, mDpDecimal, mDpDecimal));
 
         if (getSActivity().mGrant.getAuthorization().getTypeUrl().contains(Authz.StakeAuthorization.getDescriptor().getFullName())) {
             try {
                 Authz.StakeAuthorization stakeAuth = Authz.StakeAuthorization.parseFrom(getSActivity().mGrant.getAuthorization().getValue());
                 if (stakeAuth.hasMaxTokens()) {
                     BigDecimal maxAmount = new BigDecimal(stakeAuth.getMaxTokens().getAmount());
-                    if (maxAmount.compareTo(mGranterUndelegatable) <= 0) {
-                        mGranterUndelegatable = maxAmount;
+                    if (maxAmount.compareTo(mGranterDelegatable) <= 0) {
+                        mGranterDelegatable = maxAmount;
                     }
                 }
             } catch (InvalidProtocolBufferException e) {
                 e.printStackTrace();
             }
         }
-        WDp.setDpCoin(getActivity(), getBaseDao(), getSActivity().mChainConfig, getSActivity().mChainConfig.mainDenom(), mGranterUndelegatable.toPlainString(), mDenomTitle, mAvailableAmount);
+        WDp.setDpCoin(getActivity(), getBaseDao(), getSActivity().mChainConfig, getSActivity().mChainConfig.mainDenom(), mGranterDelegatable.toPlainString(), mDenomTitle, mAvailableAmount);
         onAddAmountWatcher();
     }
 
@@ -149,7 +157,7 @@ public class AuthzUndelegateStep1Fragment extends BaseFragment implements View.O
                             return;
                         }
 
-                        if (inputAmount.compareTo(mGranterUndelegatable.movePointLeft(mDpDecimal).setScale(mDpDecimal, RoundingMode.CEILING)) > 0) {
+                        if (inputAmount.compareTo(mGranterDelegatable.movePointLeft(mDpDecimal).setScale(mDpDecimal, RoundingMode.CEILING)) > 0) {
                             mAmountInput.setBackground(ContextCompat.getDrawable(getSActivity(), R.drawable.edittext_box_error));
                         } else {
                             mAmountInput.setBackground(ContextCompat.getDrawable(getSActivity(), R.drawable.edittext_box));
@@ -205,11 +213,11 @@ public class AuthzUndelegateStep1Fragment extends BaseFragment implements View.O
             mAmountInput.setText(existed.add(new BigDecimal("100")).toPlainString());
 
         } else if (v.equals(mAddHalf)) {
-            BigDecimal half = mGranterUndelegatable.movePointLeft(mDpDecimal).divide(new BigDecimal("2"), mDpDecimal, RoundingMode.DOWN);
+            BigDecimal half = mGranterDelegatable.movePointLeft(mDpDecimal).divide(new BigDecimal("2"), mDpDecimal, RoundingMode.DOWN);
             mAmountInput.setText(half.toPlainString());
 
         } else if (v.equals(mAddMax)) {
-            BigDecimal max = mGranterUndelegatable.movePointLeft(mDpDecimal).setScale(mDpDecimal, RoundingMode.DOWN);
+            BigDecimal max = mGranterDelegatable.movePointLeft(mDpDecimal).setScale(mDpDecimal, RoundingMode.DOWN);
             mAmountInput.setText(max.toPlainString());
 
         } else if (v.equals(mClearAll)) {
@@ -223,7 +231,7 @@ public class AuthzUndelegateStep1Fragment extends BaseFragment implements View.O
         try {
             BigDecimal amountTemp = new BigDecimal(mAmountInput.getText().toString().trim());
             if (amountTemp.compareTo(BigDecimal.ZERO) <= 0) return false;
-            if (amountTemp.compareTo(mGranterUndelegatable.movePointLeft(mDpDecimal).setScale(mDpDecimal, RoundingMode.CEILING)) > 0)
+            if (amountTemp.compareTo(mGranterDelegatable.movePointLeft(mDpDecimal).setScale(mDpDecimal, RoundingMode.CEILING)) > 0)
                 return false;
             Coin coin = new Coin(getSActivity().mChainConfig.mainDenom(), amountTemp.movePointRight(mDpDecimal).setScale(0).toPlainString());
             getSActivity().mAmount = coin;
@@ -246,7 +254,7 @@ public class AuthzUndelegateStep1Fragment extends BaseFragment implements View.O
         }
     }
 
-    private AuthzUndelegateActivity getSActivity() {
-        return (AuthzUndelegateActivity) getBaseActivity();
+    private AuthzDelegateActivity getSActivity() {
+        return (AuthzDelegateActivity) getBaseActivity();
     }
 }
