+++ app/src/main/java/wannabit/io/cosmostaion/fragment/txs/authz/granter/AuthzGranteeFragment.kt
@@ -1,4 +1,4 @@
-package wannabit.io.cosmostaion.fragment.txs.authz.grantee
+package wannabit.io.cosmostaion.fragment.txs.authz.granter
 
 import android.os.Bundle
 import android.view.LayoutInflater
