+++ app/src/main/res/layout/fragment_revoke_step3.xml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="utf-8"?>
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    tools:context=".fragment.txs.authz.grantee.RevokeStep3Fragment">
+
+</androidx.constraintlayout.widget.ConstraintLayout>
