+++ app/src/main/res/values-ja/strings.xml
@@ -941,6 +941,7 @@
     <string name="str_authz_commission_title">Authz コミッションをもらう。</string>
     <string name="str_authz_commission_step_0">ステーキングコミッション情報確認</string>
     <string name="str_authz_no_revoke_msg">キャンセルする権限を選択してください。</string>
+    <string name="str_authz_revoke">権限の取り消しです</string>
 
     <string name="str_select_contract_token_title">トークンの編集</string>
     <string name="str_select_contract_token">表示するトークンを選択してください</string>
