+++ app/src/main/res/values-ko/strings.xml
@@ -943,6 +943,7 @@
     <string name="str_authz_commission_title">Authz 커미션 받기</string>
     <string name="str_authz_commission_step_0">Confirm your staking commission info</string>
     <string name="str_authz_no_revoke_msg">취소할 권한을 선택하세요.</string>
+    <string name="str_authz_revoke">권한 취소</string>
 
     <string name="str_select_contract_token_title">토큰 편집</string>
     <string name="str_select_contract_token">표시할 토큰을 선택하세요</string>
